renamed all occurences of the words absentie (that i could find) to absence
