+++ src/main/kotlin/dev/avt/api/admin/AdminRouting.kt
@@ -1,8 +1,8 @@
 package dev.avt.api.admin
 
-import dev.avt.api.admin.absentie.adminAbsentieRouting
+import dev.avt.api.admin.absence.adminAbsenceRouting
 import io.ktor.server.routing.*
 
 fun Routing.adminRouting() {
-    adminAbsentieRouting()
+    adminAbsenceRouting()
 }
