--- src/main/kotlin/dev/avt/api/admin/absentie/AdminAbsentieRouting.kt
@@ -1,10 +1,0 @@
-package dev.avt.api.admin.absentie
-
-import dev.avt.api.admin.absentie.approveHours.approveHoursRoute
-import dev.avt.api.admin.absentie.readRequestedHours.readRequestedHours
-import io.ktor.server.routing.*
-
-fun Routing.adminAbsentieRouting() {
-    approveHoursRoute()
-    readRequestedHours()
-}
