+++ src/main/kotlin/dev/avt/api/person/absence/AbsenceRouting.kt
@@ -1,0 +1,10 @@
+package dev.avt.api.person.absence
+
+import dev.avt.api.person.absence.availability.readAvailabilityRoutes
+import dev.avt.api.person.absence.requestHours.requestHours
+import io.ktor.server.routing.*
+
+fun Routing.absenceRouting(){
+    readAvailabilityRoutes()
+    requestHours()
+}
