+++ src/main/kotlin/dev/avt/api/person/absence/availability/ReadAvailabilityResponse.kt
@@ -1,4 +1,4 @@
-package dev.avt.api.person.absentie.availability
+package dev.avt.api.person.absence.availability
 
 import kotlinx.serialization.Serializable
