+++ src/main/kotlin/dev/avt/api/person/absence/requestHours/RequestHoursRequest.kt
@@ -1,4 +1,4 @@
-package dev.avt.api.person.absentie.requestHours
+package dev.avt.api.person.absence.requestHours
 
 import kotlinx.serialization.Serializable
