--- src/main/kotlin/dev/avt/api/person/absentie/AbsentieRouting.kt
@@ -1,10 +1,0 @@
-package dev.avt.api.person.absentie
-
-import dev.avt.api.person.absentie.availability.readAvailabilityRoutes
-import dev.avt.api.person.absentie.requestHours.requestHours
-import io.ktor.server.routing.*
-
-fun Routing.absenceRouting(){
-    readAvailabilityRoutes()
-    requestHours()
-}
